requirements.txt
flask 
joblib 
web3 
python-dotenv
scikit-learn
imblearn
kagglehub
xgboost
seaborn
pandas
matplotlib
